refactor: Lowercase filenames
Align with Gutenberg project practices.
