+++ ReactApp/src/components/editor-load-notice.jsx
