+++ ReactApp/src/components/editor-toolbar.jsx
