+++ ReactApp/src/remote.jsx
@@ -47,7 +47,7 @@ async function initalizeRemoteEditor() {
 		const post = getPost();
 		const settings = { post };
 
-		const { default: App } = await import('./App.jsx');
+		const { default: App } = await import('./app');
 		const { createRoot, createElement, StrictMode } = window.wp.element;
 		createRoot(document.getElementById('root')).render(
 			createElement(StrictMode, null, createElement(App, settings))
